requirements.txt
google-generativeai>=0.8.0
notion-client>=2.0.0
python-dotenv>=1.0.0
docx2txt>=0.8
requests>=2.28.0
